refactor!: remove deprecated sysdig_secure_policy resource
