+++ sysdig/provider.go
@@ -130,7 +130,6 @@ func (p *SysdigProvider) Provider() *schema.Provider {
 			"sysdig_monitor_alert_group_outlier": deprecatedResourceSysdigMonitorAlertGroupOutlier(),
 			"sysdig_monitor_alert_metric":        deprecatedResourceSysdigMonitorAlertMetric(),
 			"sysdig_monitor_alert_promql":        deprecatedResourceSysdigMonitorAlertPromql(),
-			"sysdig_secure_policy":               deprecatedResourceSysdigSecurePolicy(),
 
 			"sysdig_monitor_alert_v2_change":                               resourceSysdigMonitorAlertV2Change(),
 			"sysdig_monitor_alert_v2_downtime":                             resourceSysdigMonitorAlertV2Downtime(),
